Internationalisation des messages

+++ src/main/java/com/devsecops/ensa/employee/service/EmployeeService.java
@@ -5,31 +5,27 @@ import com.devsecops.ensa.employee.repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
 @Service
 public class EmployeeService {
 
     @Autowired
     private EmployeeRepository repository;
 
-    public Employee saveEmployee(Employee employee) {
+    public Employee saveEmployee(Employee employee) throws Exception {
+        if (repository.findByEmail(employee.getEmail()).isPresent()) {
+            throw new Exception("Email already exists");
+        }
         return repository.save(employee);
-    }
-
-    public List<Employee> getAllEmployees() {
-        return repository.findAll();
     }
 
     public Employee getEmployeeById(Long id) {
         return repository.findById(id).orElse(null);
     }
 
-    public void deleteEmployee(Long id) {
+    public void deleteEmployee(Long id) throws Exception {
+        if (!repository.existsById(id)) {
+            throw new Exception("Employee not found");
+        }
         repository.deleteById(id);
-    }
-
-    public Employee updateEmployee(Employee employee) {
-        return repository.save(employee);
     }
 }
